perPage fixed

+++ src/hooks/cartQueries.js
@@ -3,6 +3,7 @@
 
 import { createData, deleteData, getData, updateData } from "@/lib/api";
 import { useQuery, useMutation } from "@tanstack/react-query";
+import { toast } from "sonner";
 
 // Fetch user's cart
 export function useCartQuery(user) {
@@ -17,12 +18,17 @@ export function useCartQuery(user) {
 }
 
 // Add item to cart
-export function useAddCartItem() {
+export function useAddCartItem(user) {
   return useMutation({
     mutationFn: async (values) => {
+      if (!user || user.role !== "customer") {
+        toast.error("Only customers/guests can add items to the cart.");
+        return;
+      }
+
       const { data } = await createData({
         url: "/cart",
-        data:values,
+        data: values,
         requiresAuth: true,
       });
       return data;
@@ -34,6 +40,10 @@ export function useAddCartItem() {
 export function useUpdateCartItem({ ...mutationOptions }) {
   return useMutation({
     mutationFn: async ({ id, quantity, address, bookingTime }) => {
+      if (!user || user.role !== "customer") {
+        toast.error("Only customers/guests can update cart items.");
+        return;
+      }
       const { data } = await updateData({
         url: `/cart/${id}`,
         data: {
@@ -52,6 +62,11 @@ export function useUpdateCartItem({ ...mutationOptions }) {
 export function useDeleteCartItem() {
   return useMutation({
     mutationFn: async (id) => {
+      if (!user || user.role !== "customer") {
+        toast.error("Only customers/guests can delete cart items.");
+        return;
+      }
+
       const { data } = await deleteData({
         url: `/cart/${id}`,
         requiresAuth: true,
